sentinelhq/portal/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
requests==2.31.0
